Move laptop to latest unstable
Citrix Workspace is now broken. Unbreak it by pulling in the missing
dependency from pinned NixOs 25.05. This is horrible, but there does not
seem to be other way around.

(Although restructure things a little bit to decouple Nixpkgs used by
Home Manager for user packages from the system ones. In theory, this
should have made it possible to allow insecure/unfree packages just for
the user profile, but it does not work, not sure why at the moment.)